# PERF: copy cached attributes on index shallow_copy