@jbrockmendel 2020-03-10
Should engine be shared?  If so, maybe access it to ensure it is in _cache before coping _cache?