@topper-123 2020-03-10
If the _cache has been populated before the call to ``_shallow_copy``, we always copy. It it hasn't, it's a question of evaluating trade-offs:
1. If we access it before copying, we risk calculating too often
2. If we do not access it before copying, we risk having to do the same calculation several times.

My hunch is that it's mostly not that important, because either the cache has already been populated, or the original index will not be reused (e.g. in a pipe, often).

But in some given cases we'd prefer either 1 or 2, depending on context, but we'd have to choose. I prefer 2., because in that case a pipe will not stop while creating ``_engine`` attributes, i..e performance will be more predictable.