@jbrockmendel 2020-03-10
Makes sense, thanks for explaining your thought process.